# src/swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Canban API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "sing up a new user",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "description": "new user credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignUpCred"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "sign up success",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "login user",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "description": "user credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLoginCred"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "access and refresh tokens",
            "schema": {
              "$ref": "#/definitions/Tokens"
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "refresh access token",
        "parameters": [
          {
            "name": "token",
            "in": "body",
            "description": "user refresh token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RefreshToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "access and refresh tokens",
            "schema": {
              "$ref": "#/definitions/Tokens"
            }
          }
        }
      }
    },
...
    "/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "description": "getting a project by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        }
      }
    },
    "/projects/create": {
      "post": {
        "tags": [
          "Projects"
        ],
        "description": "create a new project",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "new project",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewProject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "created project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        }
      }
    },
    "/projects/update/{id}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "description": "update an existing project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "project fields to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewProject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "updated project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        }
      }
    },
    "/projects/delete/{id}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "description": "project removal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "removal success",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          }
        }
      }
    },
...
    "/boards/{id}": {
      "get": {
        "tags": [
          "Boards"
        ],
        "description": "getting a board by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "board",
            "schema": {
              "$ref": "#/definitions/Board"
            }
          }
        }
      }
    },
    "/boards/create": {
      "post": {
        "tags": [
          "Boards"
        ],
        "description": "create a new board",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "new board data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewBoard"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "created board",
            "schema": {
              "$ref": "#/definitions/Boards"
            }
          }
        }
      }
    },
    "/boards/update/{id}": {
      "put": {
        "tags": [
          "Boards"
        ],
        "description": "update an existing board",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "board fields to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewBoard"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "updated board",
            "schema": {
              "$ref": "#/definitions/Boards"
            }
          }
        }
      }
    },
    "/boards/delete/{id}": {
      "delete": {
        "tags": [
          "Boards"
        ],
        "description": "board removal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "removal success",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Tokens": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYxZjIxZmM1MjM3OGNjMzQzM2E3MTkxYiIsImlhdCI6MTY0MzQzNTI0MiwiZXhwIjoxNjQzNDM3MDQyfQ.j1tQjw1tK4qQBZ7Rcfvly2J5Q0THoWc7BiTnHPFIrGc"
        },
        "refreshToken": {
          "type": "string",
          "example": "ec59ea02-4610-43e5-ba81-a9fb084e82c5"
        }
      }
    },
    "RefreshToken": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string",
          "example": "ec59ea02-4610-43e5-ba81-a9fb084e82c5"
        }
      }
    },
    "SignUpCred": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string",
          "example": "john@example.com"
        },
        "password": {
          "type": "string",
          "example": "12345678"
        },
        "name": {
          "type": "string",
          "example": "john"
        }
      }
    },
    "UserLoginCred": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string",
          "example": "john@example.com"
        },
        "password": {
          "type": "string",
          "example": "12345678"
        }
      }
    },
    "Success": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
...
    "Project": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "61e65cd062cdb8288b0e6484"
        },
        "name": {
          "type": "string",
          "example": "default project"
        },
        "description": {
          "type": "string",
          "example": "default project description"
        }
      }
    },
    "NewProject": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "default project"
        },
        "description": {
          "type": "string",
          "example": "default project description"
        }
      }
    },
    "Board": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "61e65d624327f14b2f7bf00c"
        },
        "name": {
          "type": "string",
          "example": "default board"
        },
        "projectId": {
          "type": "string",
          "example": "61e65cd062cdb8288b0e6484"
        }
      }
    },
    "Boards": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Board"
      }
    },
    "NewBoard": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "default board"
        },
        "projectId": {
          "type": "string",
          "example": "61e65cd062cdb8288b0e6484"
        }
      }
    },
...
  }
}